devel/go: actually, removed that as a whole
It's not really fair to force users into pie, even if we think it's correct.
It has some side-effects even when done the "right" way.
So we build our own version as a pie,
and offer a profile entry for users to enable if they want it by default too.