devel/rust: mark as arch=!all
Until the rust project can manage to release 'stable' versions that don't require handholding and intimate knowledge of the design just to build itself then it is not feasible to maintain. Coupled with the fact that it is basically impossible to bootstrap, the upstream rustup toolchains are useless and the general attitude towards native building instead of using their as mentioned, useless toolchains: unless someone steps up to maintain this package going forward it will not be built and will likely be removed entirely.